Extract methods for reading/writing SDC
