+++ tests/apis/wikimedia/test_structured_data_methods.py
@@ -1,5 +1,6 @@
 import pytest
 
+from flickypedia.apis.structured_data import create_license_statement
 from flickypedia.apis import UnknownWikimediaApiException, WikimediaApi
 
 
@@ -53,3 +54,66 @@ class TestAddFileCaption:
             )
 
         assert exc.value.code == "no-such-entity-link"
+
+
+class TestAddStructuredData:
+    def test_can_add_structured_data(self, wikimedia_api: WikimediaApi) -> None:
+        # This test was run against one of my Wikimedia Commons images
+        # which didn't have any SDC attached; I added the license statement
+        # and checked that it was updated as part of the process.
+        before_statements = wikimedia_api.get_structured_data(
+            filename="PrintedLibraryOfCongressSubjectHeadings.jpg"
+        )
+
+        assert "P275" not in before_statements
+
+        wikimedia_api.add_structured_data(
+            filename="PrintedLibraryOfCongressSubjectHeadings.jpg",
+            data={"claims": [create_license_statement(license_id="cc-by-2.0")]},
+            summary="Flickypedia edit (add structured data statements)",
+        )
+
+        after_statements = wikimedia_api.get_structured_data(
+            filename="PrintedLibraryOfCongressSubjectHeadings.jpg"
+        )
+
+        assert after_statements["P275"] == [
+            {
+                "id": "M138833893$3EDD9F54-EFDF-4D25-A431-F4A03486265D",
+                "mainsnak": {
+                    "datavalue": {
+                        "type": "wikibase-entityid",
+                        "value": {
+                            "entity-type": "item",
+                            "id": "Q19125117",
+                            "numeric-id": 19125117,
+                        },
+                    },
+                    "hash": "82d5b402aa0a79040483a8c9264bd484c6c13f67",
+                    "property": "P275",
+                    "snaktype": "value",
+                },
+                "rank": "normal",
+                "type": "statement",
+            }
+        ]
+
+    def test_fails_sdc_if_file_does_not_exist(
+        self, wikimedia_api: WikimediaApi
+    ) -> None:
+        with pytest.raises(UnknownWikimediaApiException) as exc:
+            wikimedia_api.add_structured_data(
+                filename="!!!.jpg",
+                data={"claims": [create_license_statement(license_id="cc-by-2.0")]},
+                summary="Flickypedia edit (add structured data statements)",
+            )
+
+        assert exc.value.code == "no-such-entity-link"
+
+    def test_throws_error_for_bad_sdc_format(self, wikimedia_api: WikimediaApi) -> None:
+        with pytest.raises(TypeError):
+            wikimedia_api.add_structured_data(
+                filename="example.jpg",
+                data=[create_license_statement(license_id="cc-by-2.0")],  # type: ignore
+                summary="Flickypedia edit (add structured data statements)",
+            )
